cleanup GQL queries
remove unused queries
add fragments
