+++ admin-ui/src/component/my-units/create-reservation/queries.ts
@@ -1,44 +1,5 @@
 import { gql } from "@apollo/client";
 
-export const RESERVATION_UNIT_QUERY = gql`
-  query reservationUnits($pk: [ID]) {
-    reservationUnits(onlyWithPermission: true, pk: $pk) {
-      edges {
-        node {
-          nameFi
-          maxPersons
-          pk
-          bufferTimeBefore
-          bufferTimeAfter
-          reservationStartInterval
-          metadataSet {
-            name
-            supportedFields
-            requiredFields
-          }
-          cancellationTerms {
-            textFi
-            nameFi
-          }
-          paymentTerms {
-            textFi
-            nameFi
-          }
-          pricingTerms {
-            textFi
-            nameFi
-          }
-          termsOfUseFi
-          serviceSpecificTerms {
-            textFi
-            nameFi
-          }
-        }
-      }
-    }
-  }
-`;
-
 export const CREATE_STAFF_RESERVATION = gql`
   mutation createStaffReservation(
     $input: ReservationStaffCreateMutationInput!
