+++ admin-ui/src/component/my-units/hooks/queries.ts
@@ -112,3 +112,51 @@ export const RESERVATION_UNITS_BY_UNIT = gql`
     }
   }
 `;
+
+// TODO bump the fragment up so we can reuse it
+// TODO this can be futher divided (for example form data, tos, base)
+const RESERVATION_UNIT_FRAGMENT = gql`
+  fragment ReservationUnitTos on ReservationUnitType {
+    nameFi
+    maxPersons
+    pk
+    bufferTimeBefore
+    bufferTimeAfter
+    reservationStartInterval
+    metadataSet {
+      name
+      supportedFields
+      requiredFields
+    }
+    cancellationTerms {
+      textFi
+      nameFi
+    }
+    paymentTerms {
+      textFi
+      nameFi
+    }
+    pricingTerms {
+      textFi
+      nameFi
+    }
+    termsOfUseFi
+    serviceSpecificTerms {
+      textFi
+      nameFi
+    }
+  }
+`;
+
+export const RESERVATION_UNIT_QUERY = gql`
+  ${RESERVATION_UNIT_FRAGMENT}
+  query reservationUnits($pk: [ID]) {
+    reservationUnits(onlyWithPermission: true, pk: $pk) {
+      edges {
+        node {
+          ...ReservationUnitTos
+        }
+      }
+    }
+  }
+`;
